Restructure repository to Terraform provider layout
