+++ internal/client/domains_test.go
@@ -1,13 +1,11 @@
 // Package domains_test contains tests for the domains package.
-package domains_test
+package client
 
 import (
 	"net/http"
 	"os"
 	"testing"
 
-	"github.com/charpand/terraform-provider-openprovider"
-	"github.com/charpand/terraform-provider-openprovider/domains"
 	"github.com/charpand/terraform-provider-openprovider/internal/testutils"
 )
 
@@ -21,15 +19,15 @@ func TestListDomains(t *testing.T) {
 		Transport: &testutils.MockTransport{RT: http.DefaultTransport},
 	}
 
-	config := openprovider.Config{
+	config := Config{
 		BaseURL:    baseURL,
 		Username:   "test",
 		Password:   "test",
 		HTTPClient: httpClient,
 	}
-	client := openprovider.NewClient(config)
+	client := NewClient(config)
 
-	resp, err := domains.List(client)
+	resp, err := List(client)
 
 	if err != nil {
 		t.Fatalf("Expected no error, got %v", err)
