+++ internal/client/get.go
@@ -1,13 +1,11 @@
-// Package domains provides functionality for working with domains.
-package domains
+// Package client provides functionality for working with domains.
+package client
 
 import (
 	"encoding/json"
 	"fmt"
 	"log"
 	"net/http"
-
-	"github.com/charpand/terraform-provider-openprovider"
 )
 
 // GetDomainResponse represents a response for a single domain.
@@ -19,7 +17,7 @@ type GetDomainResponse struct {
 // Get retrieves a single domain by ID from the Openprovider API.
 //
 // Endpoint: GET https://api.openprovider.eu/v1beta/domains/{id}
-func Get(c *openprovider.Client, id int) (*Domain, error) {
+func Get(c *Client, id int) (*Domain, error) {
 	req, err := http.NewRequest("GET", fmt.Sprintf("%s/v1beta/domains/%d", c.BaseURL, id), nil)
 	if err != nil {
 		return nil, err
