+++ internal/client/put_test.go
@@ -1,17 +1,15 @@
 // Package domains_test contains tests for the domains package.
-package domains_test
+package client
 
 import (
 	"net/http"
 	"os"
 	"testing"
 
-	"github.com/charpand/openprovider-go"
-	"github.com/charpand/openprovider-go/domains"
-	"github.com/charpand/openprovider-go/internal/testutils"
+	"github.com/charpand/terraform-provider-openprovider/internal/testutils"
 )
 
-func TestCreateDomain(t *testing.T) {
+func TestUpdateDomain(t *testing.T) {
 	baseURL := os.Getenv("TEST_API_BASE_URL")
 	if baseURL == "" {
 		baseURL = "http://localhost:4010"
@@ -21,22 +19,20 @@ func TestCreateDomain(t *testing.T) {
 		Transport: &testutils.MockTransport{RT: http.DefaultTransport},
 	}
 
-	config := openprovider.Config{
+	config := Config{
 		BaseURL:    baseURL,
 		Username:   "test",
 		Password:   "test",
 		HTTPClient: httpClient,
 	}
-	client := openprovider.NewClient(config)
+	client := NewClient(config)
 
-	// Create a test domain request
-	req := &domains.CreateDomainRequest{}
-	req.Domain.Name = "example"
-	req.Domain.Extension = "com"
-	req.OwnerHandle = "testowner"
-	req.Period = 1
+	// Update a test domain
+	req := &UpdateDomainRequest{
+		Autorenew: "on",
+	}
 
-	domain, err := domains.Create(client, req)
+	domain, err := Update(client, 123, req)
 
 	if err != nil {
 		t.Fatalf("Expected no error, got %v", err)
